+++ qlat/field.h
@@ -96,7 +96,7 @@ struct Field {
   }
   const M& get_elem(const long offset) const
   {
-    qassert(0 <= offset && offset < field.size());
+    qassert(0 <= offset && offset < (long)field.size());
     return field[offset];
   }
   //
